refactor: Move types to separate folder
